+++ docs/source/conf.py
@@ -26,6 +26,9 @@ default_role = "lua:obj"
 
 lua_ls_project_root = "example"
 lua_ls_backend = "emmylua"
+lua_ls_class_default_function_name = "__init"
+lua_ls_class_default_force_non_colon = True
+lua_ls_class_default_force_return_self = True
 
 intersphinx_mapping = {
     "python": ("https://docs.python.org/3", None),